Changed SetRotationAngle parameter to float

+++ Viewer/include/MeshModel.h
@@ -50,7 +50,7 @@ public:
 	void SetColor(const glm::vec4& color);
 	void SetTranslation(glm::vec3 direction) { translationVector = direction; }
 	void SetScaling(glm::vec3 scale)         { scaleSize         = scale; }
-	void SetRotationAngle(int angle)         { rotateAngle       = angle; }
+	void SetRotationAngle(float angle)         { rotateAngle       = angle; }
 
 	//Getters
 	std::vector<glm::vec3>& GetVerticesVector() { return vertices; }
